Replaced Makefile with make.sh
The features of Make are of no use to us here, so might as well use a
simpler shellscript. Plus it allows us to control where we run commands
from, for further foolproofing - the Makefile would've run commands from
wherever the user is and might not perform as expected, while with
make.sh we can cd to ensure we have the right path.